Break out TextTag. Rename to MARStore.
